refactor menu file
